pyproject.toml
[tool.poetry]
name = "diplomawork"
version = "0.1.0"
description = ""
authors = ["emotzart"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
p = "^1.4.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
dataclasses-json = "^0.5.7"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.dev.dependencies]
ansible-vault = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
